Prints message to indicate when a Geth node is stopping, to detect unstopped Geth nodes.

+++ integration/datagenerator/common.go
@@ -30,7 +30,7 @@ func RandomAddress() gethcommon.Address {
 }
 
 func randomUInt64() uint64 {
-	val, err := rand.Int(rand.Reader, big.NewInt(int64(math.MaxInt64)))
+	val, err := rand.Int(rand.Reader, big.NewInt(math.MaxInt64))
 	if err != nil {
 		panic(err)
 	}
